Move prefork into pyop2_utils
To simplify import dance, move prefork setup into separate module so we
can definitely import it first without accidentally hitting MPI imports.
