+++ pyop2/__init__.py
@@ -6,6 +6,10 @@ delivers performance-portability across a range of platforms:
 * GPU (CUDA and OpenCL)
 """
 
+from pyop2_utils import enable_mpi_prefork
+enable_mpi_prefork()
+del enable_mpi_prefork
+
 from op2 import *
 from version import __version__ as ver, __version_info__  # noqa: just expose
