+++ pyop2_utils/__init__.py
@@ -1,0 +1,61 @@
+# This file is part of PyOP2
+#
+# PyOP2 is Copyright (c) 2012-2015, Imperial College London and
+# others. Please see the AUTHORS file in the main source directory for
+# a full list of copyright holders.  All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#     * Redistributions of source code must retain the above copyright
+#       notice, this list of conditions and the following disclaimer.
+#     * Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#     * The name of Imperial College London or that of other
+#       contributors may not be used to endorse or promote products
+#       derived from this software without specific prior written
+#       permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTERS
+# ''AS IS'' AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDERS OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+# INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+# HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+# STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+# OF THE POSSIBILITY OF SUCH DAMAGE.
+
+from __future__ import absolute_import
+
+__all__ = ["enable_mpi_prefork"]
+
+from pyop2_utils import prefork
+
+
+initialized = False
+
+
+def enable_mpi_prefork():
+    """Start a fork server and then enable MPI."""
+    global initialized
+    import mpi4py.rc
+    mpi4py.rc.initialize = False
+    mpi4py.rc.finalize = True
+    from mpi4py import MPI
+    # Forker must be enabled before MPI is initialized, hence this
+    # dance.
+    if not initialized and MPI.Is_initialized():
+        raise RuntimeError("MPI initialized before fork server")
+
+    if not initialized:
+        prefork.enable_prefork()
+        initialized = True
+
+    if not MPI.Is_initialized():
+        MPI.Init()
