+++ tox.ini
@@ -1,6 +1,6 @@
 [flake8]
 ignore = E501,F403,E226,E402,E721,E731
-exclude = .git,__pycache__,build,.tox,dist,yacctab.py,lextab.py,doc/sphinx/source/conf.py,_version.py,pyop2/prefork.py
+exclude = .git,__pycache__,build,.tox,dist,yacctab.py,lextab.py,doc/sphinx/source/conf.py,_version.py,pyop2_utils/prefork.py
 [tox]
 envlist = py27
 [testenv]
